refactor: Issue 132 remove legacy parsers (#166)
* Remove legacy parser; black formatting

* Removed test cases related to legacy parsing; black formatting

* Removed legacy parser; black formatting

* Re-ran black wiht line length 120

* Minor typo fixes; added type hinting

* Removed make_parser; unused abstract methods

* Removed unused methods and functions

* Removed tests after methods removed

* Reverting previous changes

* Ran isort + black

* Re-ran isort + black

* Fixing docstring

* Renamed grammer.py->interpreter.py

* Removed dependency on pyparsing

* Updating pyproject.toml to remove pyparsing

* recreated lock file

* Rename test_grammar -> test_interpreter

* Updated imports

* Removed reference to pyparsing

* Correcting import statements
