+++ pyproject.toml
@@ -11,7 +11,6 @@ readme = "README.md"
 python = "^3.10"
 pydantic = "^2.7"
 sympy = "^1.12"
-pyparsing ="~3.1.2"
 qref = "0.9.0"
 
 # A list of all of the optional dependencies, some of which are included in the
