+++ src/bartiq/symbolics/ast_parser.py
@@ -14,8 +14,7 @@
 
 """AST-based expression parser.
 
-Contrary to pyparsing-based grammar from grammar.py, this module
-utilizes ast to parse an arithmetic expression as Python expresson.
+This module utilizes ast to parse an arithmetic expression as Python expresson.
 Since not all expressions that we use are correct Python expressions,
 the parser first preprocesses expressions and replaces unallowed
 syntax with carefully crafted function calls. In particular:
@@ -52,7 +51,7 @@ from functools import singledispatch, singledispatchmethod
 from typing import Callable
 from warnings import warn
 
-from .grammar import Interpreter
+from .interpreter import Interpreter
 
 _BINARY_OP_MAP = {
     ast.Mult: operator.mul,
