+++ src/bartiq/symbolics/sympy_interpreter.py
@@ -13,6 +13,8 @@
 # limitations under the License.
 import operator
 from typing import Any
+from typing import Any
+
 from sympy import (
     Function,
     Heaviside,
@@ -57,7 +59,7 @@ from sympy import prod, re, sec, sech, sin, sinh, sqrt, tan, tanh
 from sympy.codegen.cfunctions import exp2, log2, log10
 from sympy.core.numbers import S as sympy_constants
 
-from bartiq.symbolics.grammar import Interpreter, debuggable
+from .interpreter import Interpreter, debuggable
 
 WILDCARD_CHARACTER: str = "~"
